Split CLI into scripts.
Not sure if this is the best way to do it, but I think it makes sense
at least.
