+++ meica/tests/test_tedana.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python
+"""
+Tests for tedana.
+"""
+
+TEST_FILE = 'https://www.dropbox.com/s/ljeqskdmnc6si9d/zcat_ffd.nii.gz?dl=0'
